# api/pkg/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Echo Rest API
  description: A sample REST API
  termsOfService: https://example.com/terms-of-service
  license:
    name: TODO
    url: TODO
  version: 1.0.0
  contact:
    name: GDSC VIT
    url: https://github.com/GDGVIT
servers:
  - url: http://localhost:3000/v1
    description: Local development server
paths:
...
  "/user/verify":
    get:
      summary: Verify User Email
      operationId: VerifyUser
      description: Use this endpoint to verify a user's email.
      parameters:
        - in: query
          name: token
          required: true
          description: Verification Token
          schema:
            type: string
        - in: query
          name: user_pid
          required: true
          description: User PID
          schema:
            type: string
      responses:
        "200":
          description: Verify User response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Verify User Response
                  value:
                    message: User email verified successfully.
                error:
                  summary: Error Verify User Response
                  value:
                    message: User email verification failed.

  ## Login
...
  "/groups":
    get:
      operationId: GroupsGet
      description: Use this endpoint to get all groups the user is a part of
      security:
        - UserJWT: []
      responses:
        "200":
          description: Groups Get Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
              examples:
                success:
                  summary: Successful Groups Get Response
                  value:
                    message: Groups were fetched successfully.
                error:
                  summary: Groups Get Failure Response
                  value:
                    message: Groups could not be fetched.
  "/groups/{pid}":
    get:
      operationId: GroupGet
      description: Use this endpoint to get a group by id
      parameters:
        - in: path
          name: pid
          required: true
          description: Group ID
          schema:
            type: string
      security:
        - UserJWT: []
      responses:
        "200":
          description: Group Get Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              examples:
                success:
                  summary: Successful Group Get Response
                  value:
                    message: Group was fetched successfully.
                error:
                  summary: Group Get Failure Response
                  value:
                    message: Group could not be fetched.
  "/groups/create":
    post:
      operationId: GroupCreate
      description: Use this endpoint to create a group
      requestBody:
        description: Group Creation Request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupCreateRequest"
      security:
        - UserJWT: []
      responses:
        "200":
          description: Group Creation Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Group Creation Response
                  value:
                    message: Group was created successfully.
                error:
                  summary: Group Creation Failure Response
                  value:
                    message: Group could not be created.
  "/groups/edit/{pid}":
    post:
      operationId: GroupEdit
      description: Use this endpoint to edit a group
      parameters:
        - in: path
          name: pid
          required: true
          description: Group ID
          schema:
            type: string
      requestBody:
        description: Group Edit Request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupEditRequest"
      security:
        - UserJWT: []
      responses:
        "200":
          description: Group Edit Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Group Edit Response
                  value:
                    message: Group was edited successfully.
                error:
                  summary: Group Edition Failure Response
                  value:
                    message: Group could not be edited.
  "/groups/delete/{pid}":
    delete:
      operationId: GroupDelete
      description: Use this endpoint to delete a group
      parameters:
        - in: path
          name: pid
          required: true
          description: Group ID
          schema:
            type: string
      security:
        - UserJWT: []
      responses:
        "200":
          description: Group Delete Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Group Delete Response
                  value:
                    message: Group was deleted successfully.
                error:
                  summary: Group Deletion Failure Response
                  value:
                    message: Group could not be deleted.
  "/groups/add/{pid}":
    post:
      operationId: GroupAdd
      description: Use this endpoint to add users to a group
      parameters:
        - in: path
          name: pid
          required: true
          description: Group ID
          schema:
            type: string
      requestBody:
        description: Group Add Request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupAddRequest"
      security:
        - UserJWT: []
      responses:
        "200":
          description: Group Add Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Group Add Response
                  value:
                    message: User was added to group successfully.
                error:
                  summary: Group Add Failure Response
                  value:
                    message: User could not be added to group.
  "/groups/remove/{pid}":
    post:
      operationId: GroupRemove
      description: Use this endpoint to remove users from a group
      parameters:
        - in: path
          name: pid
          required: true
          description: Group ID
          schema:
            type: string
      requestBody:
        description: Group Remove Request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupRemoveRequest"
      security:
        - UserJWT: []
      responses:
        "200":
          description: Group Remove Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Group Remove Response
                  value:
                    message: User was removed from group successfully.
                error:
                  summary: Group Remove Failure Response
                  value:
                    message: User could not be removed from group.

  #################################
  #### AUTHORIZATION ENDPOINTS ####
  #################################
...
components:
  schemas:
...
    GenericMessageResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
    Group:
      type: object
      properties:
        name:
          type: string
          example: Some Group Name
        description:
          type: string
          example: Some Group Description
        users:
          type: array
          items:
            $ref: "#/components/schemas/GroupUser"
            example:
              { "id": 1, "name": "Some User Name", "email": "some@email.com" }
    GroupUser:
      type: object
      required:
        - id
        - name
        - email
        - group_role
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Some User Name
        email:
          type: string
          example: user@email.com
        group_role:
          type: string
          enum: [owner, admin, editor, viewer]
          example: admin
    GroupCreateRequest:
      type: object
      required:
        - name
        - description
        - members
      properties:
        name:
          type: string
          example: Some Group Name
        description:
          type: string
          example: Some Group Description
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupCreateRequestGroupMember"
    GroupCreateRequestGroupMember:
      type: object
      required:
        - user_pid
        - group_role
      properties:
        user_pid:
          type: string
          example: abcdef
        group_role:
          type: string
          enum: [admin, editor, viewer]
          example: admin
    GroupEditRequest:
      type: object
      properties:
        name:
          type: string
          example: Some Group Name
        description:
          type: string
          example: Some Group Description
    GroupAddRequest:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        group_role:
          type: string
          enum: [admin, editor, viewer]
          example: admin
    GroupRemoveRequest:
      type: array
      items:
        type: integer
        description: userIDs to remove from group
        example: 1
...
  securitySchemes:
    UserJWT:
      type: apiKey
      in: header
      name: Authorization